+++ NEWS.md
@@ -1,11 +1,14 @@
 # hellmer 0.1.2
 
+## New Features
+* `chat_future()` more intelligently uses `chunk_size` to determine the default number of `workers` capped at the number of CPU cores.
+
+* `$batch()` gains `progress` in addition to  `echo` and `...` which are passed to the chat call
+
 ## Lifecycle changes
 * Removed the timeout feature as it's better handled by `option(ellmer_timeout_s = 120)` in ellmer 0.1.1
 
-* Moved parameters from `chat_sequential` and `chat_future` to `$batch()` except for `chat_model` and `...`
-
-* `$batch()` gains `progress` in addition to  `echo` and `...` which are passed to the chat call
+* Moved parameters from `chat_sequential()` and `chat_future()` to `$batch()` except for `chat_model` and `...`
 
 # hellmer 0.1.1